+++ widgets/segmentdisplay/segmentdisplay_test.go
@@ -24,6 +24,9 @@ import (
 	"github.com/mum4k/termdash/internal/canvas"
 	"github.com/mum4k/termdash/internal/canvas/testcanvas"
 	"github.com/mum4k/termdash/internal/faketerm"
+	"github.com/mum4k/termdash/internal/segdisp"
+	"github.com/mum4k/termdash/internal/segdisp/dotseg"
+	"github.com/mum4k/termdash/internal/segdisp/dotseg/testdotseg"
 	"github.com/mum4k/termdash/internal/segdisp/sixteen"
 	"github.com/mum4k/termdash/internal/segdisp/sixteen/testsixteen"
 	"github.com/mum4k/termdash/terminal/terminalapi"
@@ -31,11 +34,20 @@ import (
 )
 
 // mustDrawChar draws the provided character in the area of the canvas or panics.
-func mustDrawChar(cvs *canvas.Canvas, char rune, ar image.Rectangle, opts ...sixteen.Option) {
-	d := sixteen.New()
-	testsixteen.MustSetCharacter(d, char)
+func mustDrawChar(cvs *canvas.Canvas, char rune, ar image.Rectangle, cOpts ...cell.Option) {
 	c := testcanvas.MustNew(ar)
-	testsixteen.MustDraw(d, c, opts...)
+	switch {
+	case char == '.' || char == ':':
+		d := dotseg.New()
+		testdotseg.MustSetCharacter(d, char)
+		testdotseg.MustDraw(d, c, dotseg.CellOpts(cOpts...))
+
+	default:
+		d := sixteen.New()
+		testsixteen.MustSetCharacter(d, char)
+		testsixteen.MustDraw(d, c, sixteen.CellOpts(cOpts...))
+	}
+
 	testcanvas.MustCopyTo(c, cvs)
 }
 
@@ -57,7 +69,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				GapPercent(-1),
 			},
-			canvas:     image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows),
+			canvas:     image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows),
 			wantNewErr: true,
 		},
 		{
@@ -65,12 +77,12 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				GapPercent(101),
 			},
-			canvas:     image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows),
+			canvas:     image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows),
 			wantNewErr: true,
 		},
 		{
 			desc:   "write fails on invalid GapPercent (too low)",
-			canvas: image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write(
 					[]*TextChunk{NewChunk("1")},
@@ -81,7 +93,7 @@ func TestSegmentDisplay(t *testing.T) {
 		},
 		{
 			desc:   "write fails on invalid GapPercent (too high)",
-			canvas: image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write(
 					[]*TextChunk{NewChunk("1")},
@@ -92,7 +104,7 @@ func TestSegmentDisplay(t *testing.T) {
 		},
 		{
 			desc:   "fails on area too small for a segment",
-			canvas: image.Rect(0, 0, sixteen.MinCols-1, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols-1, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("1")})
 			},
@@ -100,7 +112,7 @@ func TestSegmentDisplay(t *testing.T) {
 		},
 		{
 			desc:   "write fails without chunks",
-			canvas: image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write(nil)
 			},
@@ -108,7 +120,7 @@ func TestSegmentDisplay(t *testing.T) {
 		},
 		{
 			desc:   "write fails with an empty chunk",
-			canvas: image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("")})
 			},
@@ -116,15 +128,15 @@ func TestSegmentDisplay(t *testing.T) {
 		},
 		{
 			desc:   "write fails on unsupported characters when requested",
-			canvas: image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
-				return sd.Write([]*TextChunk{NewChunk(".", WriteErrOnUnsupported())})
+				return sd.Write([]*TextChunk{NewChunk("←", WriteErrOnUnsupported())})
 			},
 			wantUpdateErr: true,
 		},
 		{
 			desc:         "draws empty without text",
-			canvas:       image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows),
+			canvas:       image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows),
 			wantCapacity: 1,
 		},
 		{
@@ -132,7 +144,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				GapPercent(0),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols*3, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols*3, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("123")})
 			},
@@ -144,9 +156,67 @@ func TestSegmentDisplay(t *testing.T) {
 					char rune
 					area image.Rectangle
 				}{
-					{'1', image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows)},
-					{'2', image.Rect(sixteen.MinCols, 0, sixteen.MinCols*2, sixteen.MinRows)},
-					{'3', image.Rect(sixteen.MinCols*2, 0, sixteen.MinCols*3, sixteen.MinRows)},
+					{'1', image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows)},
+					{'2', image.Rect(segdisp.MinCols, 0, segdisp.MinCols*2, segdisp.MinRows)},
+					{'3', image.Rect(segdisp.MinCols*2, 0, segdisp.MinCols*3, segdisp.MinRows)},
+				} {
+					mustDrawChar(cvs, tc.char, tc.area)
+				}
+
+				testcanvas.MustApply(cvs, ft)
+				return ft
+			},
+			wantCapacity: 3,
+		},
+		{
+			desc: "uses the dot segment for a colon",
+			opts: []Option{
+				GapPercent(0),
+			},
+			canvas: image.Rect(0, 0, segdisp.MinCols*3, segdisp.MinRows),
+			update: func(sd *SegmentDisplay) error {
+				return sd.Write([]*TextChunk{NewChunk("1:3")})
+			},
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				cvs := testcanvas.MustNew(ft.Area())
+
+				for _, tc := range []struct {
+					char rune
+					area image.Rectangle
+				}{
+					{'1', image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows)},
+					{':', image.Rect(segdisp.MinCols, 0, segdisp.MinCols*2, segdisp.MinRows)},
+					{'3', image.Rect(segdisp.MinCols*2, 0, segdisp.MinCols*3, segdisp.MinRows)},
+				} {
+					mustDrawChar(cvs, tc.char, tc.area)
+				}
+
+				testcanvas.MustApply(cvs, ft)
+				return ft
+			},
+			wantCapacity: 3,
+		},
+		{
+			desc: "uses the dot segment for a dot",
+			opts: []Option{
+				GapPercent(0),
+			},
+			canvas: image.Rect(0, 0, segdisp.MinCols*3, segdisp.MinRows),
+			update: func(sd *SegmentDisplay) error {
+				return sd.Write([]*TextChunk{NewChunk("1.3")})
+			},
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				cvs := testcanvas.MustNew(ft.Area())
+
+				for _, tc := range []struct {
+					char rune
+					area image.Rectangle
+				}{
+					{'1', image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows)},
+					{'.', image.Rect(segdisp.MinCols, 0, segdisp.MinCols*2, segdisp.MinRows)},
+					{'3', image.Rect(segdisp.MinCols*2, 0, segdisp.MinCols*3, segdisp.MinRows)},
 				} {
 					mustDrawChar(cvs, tc.char, tc.area)
 				}
@@ -161,15 +231,15 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				GapPercent(0),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols*2, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols*2, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
-				return sd.Write([]*TextChunk{NewChunk(".1")})
+				return sd.Write([]*TextChunk{NewChunk("←1")})
 			},
 			want: func(size image.Point) *faketerm.Terminal {
 				ft := faketerm.MustNew(size)
 				cvs := testcanvas.MustNew(ft.Area())
 
-				mustDrawChar(cvs, '1', image.Rect(sixteen.MinCols, 0, sixteen.MinCols*2, sixteen.MinRows))
+				mustDrawChar(cvs, '1', image.Rect(segdisp.MinCols, 0, segdisp.MinCols*2, segdisp.MinRows))
 
 				testcanvas.MustApply(cvs, ft)
 				return ft
@@ -181,15 +251,15 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				GapPercent(0),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols*2, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols*2, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
-				return sd.Write([]*TextChunk{NewChunk(".1", WriteSanitize())})
+				return sd.Write([]*TextChunk{NewChunk("←1", WriteSanitize())})
 			},
 			want: func(size image.Point) *faketerm.Terminal {
 				ft := faketerm.MustNew(size)
 				cvs := testcanvas.MustNew(ft.Area())
 
-				mustDrawChar(cvs, '1', image.Rect(sixteen.MinCols, 0, sixteen.MinCols*2, sixteen.MinRows))
+				mustDrawChar(cvs, '1', image.Rect(segdisp.MinCols, 0, segdisp.MinCols*2, segdisp.MinRows))
 
 				testcanvas.MustApply(cvs, ft)
 				return ft
@@ -198,7 +268,7 @@ func TestSegmentDisplay(t *testing.T) {
 		},
 		{
 			desc:   "aligns segment vertical middle by default",
-			canvas: image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows+2),
+			canvas: image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows+2),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("1")})
 			},
@@ -206,7 +276,7 @@ func TestSegmentDisplay(t *testing.T) {
 				ft := faketerm.MustNew(size)
 				cvs := testcanvas.MustNew(ft.Area())
 
-				mustDrawChar(cvs, '1', image.Rect(0, 1, sixteen.MinCols, sixteen.MinRows+1))
+				mustDrawChar(cvs, '1', image.Rect(0, 1, segdisp.MinCols, segdisp.MinRows+1))
 
 				testcanvas.MustApply(cvs, ft)
 				return ft
@@ -215,7 +285,7 @@ func TestSegmentDisplay(t *testing.T) {
 		},
 		{
 			desc:   "subsequent calls to write overwrite previous text",
-			canvas: image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows+2),
+			canvas: image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows+2),
 			update: func(sd *SegmentDisplay) error {
 				if err := sd.Write([]*TextChunk{NewChunk("123")}); err != nil {
 					return err
@@ -226,7 +296,7 @@ func TestSegmentDisplay(t *testing.T) {
 				ft := faketerm.MustNew(size)
 				cvs := testcanvas.MustNew(ft.Area())
 
-				mustDrawChar(cvs, '4', image.Rect(0, 1, sixteen.MinCols, sixteen.MinRows+1))
+				mustDrawChar(cvs, '4', image.Rect(0, 1, segdisp.MinCols, segdisp.MinRows+1))
 
 				testcanvas.MustApply(cvs, ft)
 				return ft
@@ -238,7 +308,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				GapPercent(0),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols*2, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols*2, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write(
 					[]*TextChunk{
@@ -258,19 +328,15 @@ func TestSegmentDisplay(t *testing.T) {
 
 				mustDrawChar(
 					cvs, '1',
-					image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows),
-					sixteen.CellOpts(
-						cell.FgColor(cell.ColorRed),
-						cell.BgColor(cell.ColorBlue),
-					),
+					image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows),
+					cell.FgColor(cell.ColorRed),
+					cell.BgColor(cell.ColorBlue),
 				)
 				mustDrawChar(
 					cvs, '2',
-					image.Rect(sixteen.MinCols, 0, sixteen.MinCols*2, sixteen.MinRows),
-					sixteen.CellOpts(
-						cell.FgColor(cell.ColorGreen),
-						cell.BgColor(cell.ColorYellow),
-					),
+					image.Rect(segdisp.MinCols, 0, segdisp.MinCols*2, segdisp.MinRows),
+					cell.FgColor(cell.ColorGreen),
+					cell.BgColor(cell.ColorYellow),
 				)
 
 				testcanvas.MustApply(cvs, ft)
@@ -283,7 +349,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				MaximizeDisplayedText(),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols*3, sixteen.MinRows+2),
+			canvas: image.Rect(0, 0, segdisp.MinCols*3, segdisp.MinRows+2),
 			update: func(sd *SegmentDisplay) error {
 				if err := sd.Write([]*TextChunk{NewChunk("123")}); err != nil {
 					return err
@@ -299,7 +365,7 @@ func TestSegmentDisplay(t *testing.T) {
 				GapPercent(0),
 				MaximizeDisplayedText(),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols*3, sixteen.MinRows+2),
+			canvas: image.Rect(0, 0, segdisp.MinCols*3, segdisp.MinRows+2),
 			update: func(sd *SegmentDisplay) error {
 				if err := sd.Write([]*TextChunk{NewChunk("123")}); err != nil {
 					return err
@@ -314,7 +380,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				MaximizeSegmentHeight(),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols*3, sixteen.MinRows+2),
+			canvas: image.Rect(0, 0, segdisp.MinCols*3, segdisp.MinRows+2),
 			update: func(sd *SegmentDisplay) error {
 				if err := sd.Write([]*TextChunk{NewChunk("123")}); err != nil {
 					return err
@@ -330,7 +396,7 @@ func TestSegmentDisplay(t *testing.T) {
 				GapPercent(0),
 				MaximizeSegmentHeight(),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols*3, sixteen.MinRows+2),
+			canvas: image.Rect(0, 0, segdisp.MinCols*3, segdisp.MinRows+2),
 			update: func(sd *SegmentDisplay) error {
 				if err := sd.Write([]*TextChunk{NewChunk("123")}); err != nil {
 					return err
@@ -342,7 +408,7 @@ func TestSegmentDisplay(t *testing.T) {
 		},
 		{
 			desc:   "reset resets provided cell options",
-			canvas: image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				if err := sd.Write(
 					[]*TextChunk{
@@ -360,7 +426,7 @@ func TestSegmentDisplay(t *testing.T) {
 				ft := faketerm.MustNew(size)
 				cvs := testcanvas.MustNew(ft.Area())
 
-				mustDrawChar(cvs, '1', image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows))
+				mustDrawChar(cvs, '1', image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows))
 
 				testcanvas.MustApply(cvs, ft)
 				return ft
@@ -372,7 +438,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				AlignVertical(align.VerticalMiddle),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows+2),
+			canvas: image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows+2),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("1")})
 			},
@@ -380,7 +446,7 @@ func TestSegmentDisplay(t *testing.T) {
 				ft := faketerm.MustNew(size)
 				cvs := testcanvas.MustNew(ft.Area())
 
-				mustDrawChar(cvs, '1', image.Rect(0, 1, sixteen.MinCols, sixteen.MinRows+1))
+				mustDrawChar(cvs, '1', image.Rect(0, 1, segdisp.MinCols, segdisp.MinRows+1))
 
 				testcanvas.MustApply(cvs, ft)
 				return ft
@@ -392,7 +458,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				AlignVertical(align.VerticalTop),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows+2),
+			canvas: image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows+2),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("1")})
 			},
@@ -400,7 +466,7 @@ func TestSegmentDisplay(t *testing.T) {
 				ft := faketerm.MustNew(size)
 				cvs := testcanvas.MustNew(ft.Area())
 
-				mustDrawChar(cvs, '1', image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows))
+				mustDrawChar(cvs, '1', image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows))
 
 				testcanvas.MustApply(cvs, ft)
 				return ft
@@ -412,7 +478,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				AlignVertical(align.VerticalBottom),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows+2),
+			canvas: image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows+2),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write(
 					[]*TextChunk{NewChunk("1")},
@@ -423,7 +489,7 @@ func TestSegmentDisplay(t *testing.T) {
 				ft := faketerm.MustNew(size)
 				cvs := testcanvas.MustNew(ft.Area())
 
-				mustDrawChar(cvs, '1', image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows))
+				mustDrawChar(cvs, '1', image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows))
 
 				testcanvas.MustApply(cvs, ft)
 				return ft
@@ -435,7 +501,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				AlignVertical(align.VerticalBottom),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows+2),
+			canvas: image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows+2),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("1")})
 			},
@@ -443,7 +509,7 @@ func TestSegmentDisplay(t *testing.T) {
 				ft := faketerm.MustNew(size)
 				cvs := testcanvas.MustNew(ft.Area())
 
-				mustDrawChar(cvs, '1', image.Rect(0, 2, sixteen.MinCols, sixteen.MinRows+2))
+				mustDrawChar(cvs, '1', image.Rect(0, 2, segdisp.MinCols, segdisp.MinRows+2))
 
 				testcanvas.MustApply(cvs, ft)
 				return ft
@@ -452,7 +518,7 @@ func TestSegmentDisplay(t *testing.T) {
 		},
 		{
 			desc:   "aligns segment horizontal center by default",
-			canvas: image.Rect(0, 0, sixteen.MinCols+2, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols+2, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("8")})
 			},
@@ -460,7 +526,7 @@ func TestSegmentDisplay(t *testing.T) {
 				ft := faketerm.MustNew(size)
 				cvs := testcanvas.MustNew(ft.Area())
 
-				mustDrawChar(cvs, '8', image.Rect(1, 0, sixteen.MinCols+1, sixteen.MinRows))
+				mustDrawChar(cvs, '8', image.Rect(1, 0, segdisp.MinCols+1, segdisp.MinRows))
 
 				testcanvas.MustApply(cvs, ft)
 				return ft
@@ -472,7 +538,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				AlignHorizontal(align.HorizontalCenter),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols+2, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols+2, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("8")})
 			},
@@ -480,7 +546,7 @@ func TestSegmentDisplay(t *testing.T) {
 				ft := faketerm.MustNew(size)
 				cvs := testcanvas.MustNew(ft.Area())
 
-				mustDrawChar(cvs, '8', image.Rect(1, 0, sixteen.MinCols+1, sixteen.MinRows))
+				mustDrawChar(cvs, '8', image.Rect(1, 0, segdisp.MinCols+1, segdisp.MinRows))
 
 				testcanvas.MustApply(cvs, ft)
 				return ft
@@ -492,7 +558,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				AlignHorizontal(align.HorizontalLeft),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols+2, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols+2, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("8")})
 			},
@@ -500,7 +566,7 @@ func TestSegmentDisplay(t *testing.T) {
 				ft := faketerm.MustNew(size)
 				cvs := testcanvas.MustNew(ft.Area())
 
-				mustDrawChar(cvs, '8', image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows))
+				mustDrawChar(cvs, '8', image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows))
 
 				testcanvas.MustApply(cvs, ft)
 				return ft
@@ -512,7 +578,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				AlignHorizontal(align.HorizontalRight),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols+2, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols+2, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("8")})
 			},
@@ -520,7 +586,7 @@ func TestSegmentDisplay(t *testing.T) {
 				ft := faketerm.MustNew(size)
 				cvs := testcanvas.MustNew(ft.Area())
 
-				mustDrawChar(cvs, '8', image.Rect(2, 0, sixteen.MinCols+2, sixteen.MinRows))
+				mustDrawChar(cvs, '8', image.Rect(2, 0, segdisp.MinCols+2, segdisp.MinRows))
 
 				testcanvas.MustApply(cvs, ft)
 				return ft
@@ -533,7 +599,7 @@ func TestSegmentDisplay(t *testing.T) {
 				MaximizeSegmentHeight(),
 				GapPercent(0),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols*2, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols*2, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("123")})
 			},
@@ -545,8 +611,8 @@ func TestSegmentDisplay(t *testing.T) {
 					char rune
 					area image.Rectangle
 				}{
-					{'1', image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows)},
-					{'2', image.Rect(sixteen.MinCols, 0, sixteen.MinCols*2, sixteen.MinRows)},
+					{'1', image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows)},
+					{'2', image.Rect(segdisp.MinCols, 0, segdisp.MinCols*2, segdisp.MinRows)},
 				} {
 					mustDrawChar(cvs, tc.char, tc.area)
 				}
@@ -561,7 +627,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				GapPercent(0),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols*3, sixteen.MinRows*4),
+			canvas: image.Rect(0, 0, segdisp.MinCols*3, segdisp.MinRows*4),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("123")})
 			},
@@ -590,7 +656,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				GapPercent(0),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols*3, sixteen.MinRows*4),
+			canvas: image.Rect(0, 0, segdisp.MinCols*3, segdisp.MinRows*4),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("1234")})
 			},
@@ -620,7 +686,7 @@ func TestSegmentDisplay(t *testing.T) {
 				MaximizeDisplayedText(),
 				GapPercent(0),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols*3, sixteen.MinRows*4),
+			canvas: image.Rect(0, 0, segdisp.MinCols*3, segdisp.MinRows*4),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("123")})
 			},
@@ -646,7 +712,7 @@ func TestSegmentDisplay(t *testing.T) {
 		},
 		{
 			desc:   "draws multiple segments with a gap by default",
-			canvas: image.Rect(0, 0, sixteen.MinCols*3+2, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols*3+2, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("123")})
 			},
@@ -658,9 +724,9 @@ func TestSegmentDisplay(t *testing.T) {
 					char rune
 					area image.Rectangle
 				}{
-					{'1', image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows)},
-					{'2', image.Rect(sixteen.MinCols+1, 0, sixteen.MinCols*2+1, sixteen.MinRows)},
-					{'3', image.Rect(sixteen.MinCols*2+2, 0, sixteen.MinCols*3+2, sixteen.MinRows)},
+					{'1', image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows)},
+					{'2', image.Rect(segdisp.MinCols+1, 0, segdisp.MinCols*2+1, segdisp.MinRows)},
+					{'3', image.Rect(segdisp.MinCols*2+2, 0, segdisp.MinCols*3+2, segdisp.MinRows)},
 				} {
 					mustDrawChar(cvs, tc.char, tc.area)
 				}
@@ -675,7 +741,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				GapPercent(20),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols*3+2, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols*3+2, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("123")})
 			},
@@ -687,9 +753,9 @@ func TestSegmentDisplay(t *testing.T) {
 					char rune
 					area image.Rectangle
 				}{
-					{'1', image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows)},
-					{'2', image.Rect(sixteen.MinCols+1, 0, sixteen.MinCols*2+1, sixteen.MinRows)},
-					{'3', image.Rect(sixteen.MinCols*2+2, 0, sixteen.MinCols*3+2, sixteen.MinRows)},
+					{'1', image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows)},
+					{'2', image.Rect(segdisp.MinCols+1, 0, segdisp.MinCols*2+1, segdisp.MinRows)},
+					{'3', image.Rect(segdisp.MinCols*2+2, 0, segdisp.MinCols*3+2, segdisp.MinRows)},
 				} {
 					mustDrawChar(cvs, tc.char, tc.area)
 				}
@@ -704,7 +770,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				GapPercent(40),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols*3+2, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols*3+2, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("123")})
 			},
@@ -732,7 +798,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				GapPercent(20),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols*3+2, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols*3+2, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("8888")})
 			},
@@ -744,9 +810,9 @@ func TestSegmentDisplay(t *testing.T) {
 					char rune
 					area image.Rectangle
 				}{
-					{'8', image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows)},
-					{'8', image.Rect(sixteen.MinCols+1, 0, sixteen.MinCols*2+1, sixteen.MinRows)},
-					{'8', image.Rect(sixteen.MinCols*2+2, 0, sixteen.MinCols*3+2, sixteen.MinRows)},
+					{'8', image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows)},
+					{'8', image.Rect(segdisp.MinCols+1, 0, segdisp.MinCols*2+1, segdisp.MinRows)},
+					{'8', image.Rect(segdisp.MinCols*2+2, 0, segdisp.MinCols*3+2, segdisp.MinRows)},
 				} {
 					mustDrawChar(cvs, tc.char, tc.area)
 				}
@@ -761,7 +827,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				GapPercent(20),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols*4+2, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols*4+2, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("8888")})
 			},
@@ -791,7 +857,7 @@ func TestSegmentDisplay(t *testing.T) {
 				MaximizeSegmentHeight(),
 				GapPercent(20),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols*5, sixteen.MinRows*2),
+			canvas: image.Rect(0, 0, segdisp.MinCols*5, segdisp.MinRows*2),
 			update: func(sd *SegmentDisplay) error {
 				return sd.Write([]*TextChunk{NewChunk("123")})
 			},
@@ -819,7 +885,7 @@ func TestSegmentDisplay(t *testing.T) {
 			opts: []Option{
 				GapPercent(0),
 			},
-			canvas: image.Rect(0, 0, sixteen.MinCols*3, sixteen.MinRows),
+			canvas: image.Rect(0, 0, segdisp.MinCols*3, segdisp.MinRows),
 			update: func(sd *SegmentDisplay) error {
 				chunks := []*TextChunk{NewChunk("123")}
 				if err := sd.Write(chunks); err != nil {
@@ -837,9 +903,9 @@ func TestSegmentDisplay(t *testing.T) {
 					char rune
 					area image.Rectangle
 				}{
-					{'1', image.Rect(0, 0, sixteen.MinCols, sixteen.MinRows)},
-					{'2', image.Rect(sixteen.MinCols, 0, sixteen.MinCols*2, sixteen.MinRows)},
-					{'3', image.Rect(sixteen.MinCols*2, 0, sixteen.MinCols*3, sixteen.MinRows)},
+					{'1', image.Rect(0, 0, segdisp.MinCols, segdisp.MinRows)},
+					{'2', image.Rect(segdisp.MinCols, 0, segdisp.MinCols*2, segdisp.MinRows)},
+					{'3', image.Rect(segdisp.MinCols*2, 0, segdisp.MinCols*3, segdisp.MinRows)},
 				} {
 					mustDrawChar(cvs, tc.char, tc.area)
 				}
@@ -938,7 +1004,7 @@ func TestOptions(t *testing.T) {
 	}
 	got := sd.Options()
 	want := widgetapi.Options{
-		MinimumSize:  image.Point{sixteen.MinCols, sixteen.MinRows},
+		MinimumSize:  image.Point{segdisp.MinCols, segdisp.MinRows},
 		WantKeyboard: widgetapi.KeyScopeNone,
 		WantMouse:    widgetapi.MouseScopeNone,
 	}